Book Model has been revamped and testcase were amended to match the modifications.
Book Model has been modified to welcome new type of data. Now, the model will capture the name of the author, the date of publication and the ISBN number. All these values must be provided for a valid input of data.

Testcases were modified to test the validity of this new model. One testcase was also created to test the '/api/books/{id}/chapters/' link.