A Quote Model has been created to store all the character's quote. An API endpoint as well as a variety of testcases were added.
The Quote Model class was implemented. It included the automatic id of the class, the text of the quote and two foreign keys: Character's id and Books's id.

An API endpoint was created and can be accessed via '/api/quotes/' or '/api/quotes/{id}/'.

Finally, a variety of testcases were added to assure the validity of the new implementation.