A Permission system has been implemented to allow certain endpoints to be only accessible by admin users.
Each ViewSet classes has been modified. A permission object was created to store each permission per action. The 'get_permissions' function at the end of each ViewSet class regulate the permissions. The 'create', 'partial_update', 'update' and 'destroy' had to be overwritten to instaure the permission functions.

New testcases were added. All the admin endpoints have two test: One with an admin input and one with a normal user input. This checks the validity of the permissions functions.
To do so, an extra user was added at the beginning of the testcase. When dealing with a no_permission test, self.client uses the force_authenticate function to connect to the normal user to check that this user has no access to this endpoint.